requirements.txt
beautifulsoup4==4.6.0
bleach==2.1.2
certifi==2017.11.5
chardet==3.0.4
defusedxml==0.5.0
Django==1.10.8
django-allauth==0.34.0
django-anymail==1.3
django-app-namespace-template-loader==0.4.1
django-appconf==1.0.1
django-avatar==4.1.0
django-badgify==0.2.0
django-chroniker==1.0.10
django-comments-xtd==2.0.9
django-compressor==2.2
django-contrib-comments==1.8.0
django-countries==5.0
django-crontab==0.7.1
django-el-pagination==3.1.0
django-extensions==1.9.8
django-grappelli==2.9.1
django-guardian==1.4.9
django-haystack==2.6.1
django-js-asset==0.1.1
django-markdown2==0.3.0
django-markdownify==0.3.0
django-markdownx==2.0.22
django-meta==1.4
django-newsletter==0.6
django-pagedown==1.0.4
django-pure-pagination==0.3.0
django-robots==3.0
django-sorl-cropping==0.1.0
django-taggit==0.22.2
django-user-agents==0.3.2
django-webtest==1.9.2
django-widget-tweaks==1.4.1
djangorestframework==3.6.4
docutils==0.14
elasticsearch==6.0.0
filebrowser-safe==0.4.7
future==0.16.0
grappelli-safe==0.4.7
html5lib==1.0.1
idna==2.6
ipaddress==1.0.19
ldif3==3.1.1
Markdown==2.6.11
markdown2==2.3.5
oauthlib==2.0.6
olefile==0.44
pilkit==2.0
Pillow==4.3.0
pkg-resources==0.0.0
psutil==5.4.2
psycopg2==2.7.3.2
pyparsing==2.2.0
python-card-me==0.9.3
python-dateutil==2.6.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2017.3
PyYAML==3.12
rcssmin==1.0.6
regex==2018.1.10
requests==2.18.4
requests-oauthlib==0.8.0
rjsmin==1.0.12
six==1.11.0
sorl-thumbnail==12.4.1
surlex==0.2.0
toposort==1.5
typing==3.6.2
tzlocal==1.5.1
ua-parser==0.7.3
unicodecsv==0.14.1
urllib3==1.22
user-agents==1.1.0
waitress==1.1.0
webencodings==0.5.1
WebOb==1.7.4
WebTest==2.0.29
